ipa-epn: Provides expiration datetime object & days count in epn template
We have internationals users and we want to format in their native language.
Here we pass the raw datimeobject and user can now use strftime in jinja
template to format depending on user language.
This PR helps in situation where their is one instance of IPA by country
one node in country X can be configured with a date format X and node in
country Y can be confiugred with a date format Y.

Maybe a more general solution is to add a country attribute and make it
visibile to template so it can dynamically switch used template per user
in the single node case.

No related issue opened.